+++ packages/typespec-rust/test/spector/client/naming/tests/naming_client.rs
@@ -58,7 +58,7 @@ async fn parameter() {
 #[tokio::test]
 async fn request() {
     let client = NamingClient::with_no_credential("http://localhost:3000", None).unwrap();
-    client.request("true", None).await.unwrap();
+    client.request("true".to_string(), None).await.unwrap();
 }
 
 #[tokio::test]