+++ packages/typespec-rust/test/spector/parameters/spread/tests/spread_alias_client_test.rs
@@ -9,7 +9,7 @@ async fn spread_as_request_body() {
     let client = SpreadClient::with_no_credential("http://localhost:3000", None).unwrap();
     client
         .get_spread_alias_client()
-        .spread_as_request_body("foo", None)
+        .spread_as_request_body("foo".to_string(), None)
         .await
         .unwrap();
 }
@@ -19,7 +19,7 @@ async fn spread_as_request_parameter() {
     let client = SpreadClient::with_no_credential("http://localhost:3000", None).unwrap();
     client
         .get_spread_alias_client()
-        .spread_as_request_parameter("1", "bar", "foo", None)
+        .spread_as_request_parameter("1", "bar".to_string(), "foo".to_string(), None)
         .await
         .unwrap();
 }
@@ -29,7 +29,7 @@ async fn spread_parameter_with_inner_alias() {
     let client = SpreadClient::with_no_credential("http://localhost:3000", None).unwrap();
     client
         .get_spread_alias_client()
-        .spread_parameter_with_inner_alias("1", "foo", 1, "bar", None)
+        .spread_parameter_with_inner_alias("1", "foo".to_string(), 1, "bar".to_string(), None)
         .await
         .unwrap();
 }
@@ -39,7 +39,7 @@ async fn spread_parameter_with_inner_model() {
     let client = SpreadClient::with_no_credential("http://localhost:3000", None).unwrap();
     client
         .get_spread_alias_client()
-        .spread_parameter_with_inner_model("1", "foo", "bar", None)
+        .spread_parameter_with_inner_model("1", "foo".to_string(), "bar".to_string(), None)
         .await
         .unwrap();
 }
@@ -51,8 +51,8 @@ async fn spread_with_multiple_parameters() {
         .get_spread_alias_client()
         .spread_with_multiple_parameters(
             "1",
-            "bar",
-            "foo",
+            "bar".to_string(),
+            "foo".to_string(),
             vec![1, 2],
             Some(SpreadAliasClientSpreadWithMultipleParametersOptions {
                 optional_int: Some(1),